requirements.txt
prometheus_client==0.16.0
psutil==5.9.4
mysql-connector-python==8.0.33
python-dotenv